Cleanup hierarchy
