+++ src/Features/CSharp/Portable/SignatureHelp/GenericNameSignatureHelpProvider_NamedType.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 
 namespace Microsoft.CodeAnalysis.CSharp.SignatureHelp;
 
-internal partial class GenericNameSignatureHelpProvider
+internal partial class AbstractGenericNameSignatureHelpProvider
 {
     private static IList<SymbolDisplayPart> GetPreambleParts(
         INamedTypeSymbol namedType,
